cleanup; removed sql injection vulnerability;
