+++ Cargo.toml
@@ -4,9 +4,11 @@ version = "0.1.0"
 edition = "2021"
 
 [workspace]
-members = ["crates/monitor_events", "crates/l2-micro-service"]
+members = ["crates/monitor_events", "crates/l2-micro-service", "crates/common"]
 
 [workspace.dependencies]
+common = { path = "crates/common" }
+
 config = "0.14.0"
 rocket = { version = "0.5.0", features = ["json"] }
 tokio-postgres = "0.7.10"
