+++ crates/common/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "common"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+eyre.workspace = true
