+++ crates/l2-micro-service/Cargo.toml
@@ -13,6 +13,7 @@ eyre.workspace = true
 tokio.workspace = true
 serde.workspace = true
 serde_json.workspace = true
+common.workspace = true
 
 [[bin]]
 name = "l2-micro-service"
