+++ crates/monitor_events/Cargo.toml
@@ -17,3 +17,4 @@ tokio-postgres.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 reqwest.workspace = true
+common.workspace = true
